refactor: move files into subfolders
